refactor(runtime): move LifecycleFlags to runtime-html
