+++ packages/__tests__/3-runtime-html/controller.host-sharing.integration.spec.ts
@@ -1,5 +1,5 @@
-import { LifecycleFlags } from '@aurelia/runtime';
 import {
+  LifecycleFlags,
   Aurelia,
   customElement,
   ICustomElementController,
