+++ packages/__tests__/3-runtime-html/enhance.spec.ts
@@ -1,8 +1,8 @@
 // This is to test for some intrinsic properties of enhance which is otherwise difficult to test in Data-driven tests parallel to `.app`
 import { BrowserPlatform } from '@aurelia/platform-browser';
 import { Constructable, DI, IContainer, Registration } from '@aurelia/kernel';
-import { LifecycleFlags } from '@aurelia/runtime';
 import {
+  LifecycleFlags,
   CustomElement,
   ICustomElementViewModel,
   IPlatform,
