+++ packages/router-lite/src/resources/viewport.ts
@@ -1,6 +1,5 @@
 import { ILogger } from '@aurelia/kernel';
-import { LifecycleFlags } from '@aurelia/runtime';
-import { bindable, customElement, ICustomElementViewModel, IHydratedController, ICustomElementController, ICompiledCustomElementController } from '@aurelia/runtime-html';
+import { LifecycleFlags, bindable, customElement, ICustomElementViewModel, IHydratedController, ICustomElementController, ICompiledCustomElementController } from '@aurelia/runtime-html';
 
 import { ViewportAgent } from '../viewport-agent';
 import { IRouteContext } from '../route-context';
