+++ packages/router-lite/src/viewport-agent.ts
@@ -1,8 +1,7 @@
 // No-fallthrough disabled due to large numbers of false positives
 /* eslint-disable no-fallthrough */
 import { ILogger, onResolve } from '@aurelia/kernel';
-import { LifecycleFlags } from '@aurelia/runtime';
-import { IHydratedController, ICustomElementController, Controller } from '@aurelia/runtime-html';
+import { LifecycleFlags, IHydratedController, ICustomElementController, Controller } from '@aurelia/runtime-html';
 
 import { IViewport } from './resources/viewport';
 import { ComponentAgent } from './component-agent';
