+++ packages/runtime-html/src/index.ts
@@ -1,8 +1,4 @@
 export {
-  // todo: only exception for now for hmr, remove
-  LifecycleFlags,
-} from '@aurelia/runtime';
-export {
   bindable,
   Bindable,
   BindableDefinition,
@@ -429,6 +425,7 @@ export {
   ViewModelKind,
   HooksDefinition,
   State,
+  LifecycleFlags,
   type ControllerVisitor,
   type IViewModel,
   IController,
