+++ packages/runtime-html/src/resources/custom-elements/au-render.ts
@@ -1,12 +1,11 @@
 import { Constructable, nextId, onResolve } from '@aurelia/kernel';
-import { LifecycleFlags } from '@aurelia/runtime';
 import { createElement, RenderPlan } from '../../create-element';
 import { HydrateElementInstruction, IInstruction } from '../../renderer';
 import { IPlatform } from '../../platform';
 import { IViewFactory } from '../../templating/view';
 import { CustomElement, customElement, CustomElementDefinition } from '../custom-element';
 import { bindable } from '../../bindable';
-import { ControllerVisitor, ICustomElementController, ICustomElementViewModel, IHydratedController, IHydratedParentController, IHydrationContext, ISyntheticView } from '../../templating/controller';
+import { LifecycleFlags, ControllerVisitor, ICustomElementController, ICustomElementViewModel, IHydratedController, IHydratedParentController, IHydrationContext, ISyntheticView } from '../../templating/controller';
 import { IRendering } from '../../templating/rendering';
 import { isPromise, isString } from '../../utilities';
 import { BindingMode } from '../../binding/interfaces-bindings';
